+++ R/cast_variables.R
@@ -3,6 +3,7 @@
 #' Convenience function to sort two strings and paste them together
 #' @param string1 first string
 #' @param string2 second string
+#' @keywords internal
 sort_then_join <- function(string1, string2) {
     paste0(sort(c(string1, string2)), collapse = "")
 }
@@ -12,6 +13,7 @@ sort_then_join <- function(string1, string2) {
 #'
 #' Convenience function to put all classes an object has into one string
 #' @param x an object
+#' @keywords internal
 class_merge <- function(x) {
     paste(class(x), collapse = "_")
 }
@@ -30,6 +32,7 @@ get_message <- function(colname, whichdat, totype) {
 #' @param colin column to cast
 #' @param colname name of vector
 #' @param whichdat whether base or compare is being casted (used for messages)
+#' @keywords internal
 get_casted_vector <- function(colin, colname, whichdat) {
     if (class(colin) == "factor") {
         get_message(colname, whichdat, "character")
@@ -53,6 +56,7 @@ get_casted_vector <- function(colin, colname, whichdat) {
 #' @param df dataset to be casted
 #' @param columns columns to be casted
 #' @param whichdat whether base or compare is being casted (used for messages)
+#' @keywords internal
 get_casted_dataset <- function(df, columns, whichdat) {
     for (col in columns) {
         df[[col]] <- get_casted_vector(df[[col]], col, whichdat)
@@ -74,6 +78,7 @@ get_casted_dataset <- function(df, columns, whichdat) {
 #' @param ignore_vars Variables not to be considered for casting
 #' @param cast_integers Logical - Whether integers should be cased to double when compared to doubles
 #' @param cast_factors Logical - Whether characters should be casted to characters when compared to characters
+#' @keywords internal
 cast_variables <- function(
         BASE,
         COMPARE,
@@ -122,7 +127,6 @@ cast_variables <- function(
         "BASE" = BASE,
         "COMPARE" = COMPARE
     )
-
 
     if (nrow(cast_columns) == 0) {
         return(DATASETS)